switch amount over sellAmount/buyAmount, minor refactoring
